+++ app/components/CardsSession.tsx
@@ -1,47 +1,46 @@
 import Link from "next/link";
 import React from "react";
+import Image from "next/image";
 
-const Card = ({ title, text, icon = "", url = "#" }) => {
+export const CardsSession = ({ props }) => {
+  const { title, cards } = props;
+  return (
+    <section>
+      <div className="flex-1 container mx-auto p-10 text-center">
+        <h3 className="font-bold text-3xl text-ameciclo py-8 w-1/2 mx-auto">
+          {title}
+        </h3>
+        <div className="grid grid-cols-1 lg:grid-cols-3 grid-rows-4 sm:grid-rows-2 gap-8 grid-flow-row">
+          {cards.map((card: any) => (
+            <Card props = {{...card}} />
+          ))}
+        </div>
+      </div>
+    </section>
+  );
+};
+
+const Card = ({ props }) => {
+  const {title, description, src = "", url= "#" } = props
   return (
     <Link href={url} target="_blank">
       <div className="bg-white w-full rounded-lg flex items-center justify-center text-ameciclo p-10">
         <div className="flex flex-col text-center">
-          {icon != "" && (
-            <img
-              src={`http://localhost:3300${icon}`}
+          {src != "" && (
+            <Image
               className="h-32 fill-current"
+              src={src}
+              alt={title}
+              height={100}
+              width={100}
             />
           )}
           <h2 className="text-2xl font-bold text-ameciclo uppercase tracking-wider my-2">
             {title}
           </h2>
-          <p className="text-base font-medium">{text}</p>
+          <p className="text-base font-medium">{description}</p>
         </div>
       </div>
     </Link>
-  );
-};
-
-export const CardSession = ({ props }) => {
-  const { title, grids } = props;
-  return (
-    <section>
-      <div className="flex-1 container mx-auto p-10 text-center">
-        <h3 className="font-bold text-3xl text-ameciclo py-8 w-1/2 mx-auto">
-          {" "}
-          {title}
-        </h3>
-        <div className="grid grid-cols-1 lg:grid-cols-3 grid-rows-4 sm:grid-rows-2 gap-8 grid-flow-row">
-          {grids?.map((grid: any) => (
-            <Card
-              title={grid.title}
-              text={grid.text}
-              icon={grid.icon}
-              url={grid.url}
-            />
-          ))}
-        </div>
-      </div>
-    </section>
   );
 };
