+++ app/ideciclo/components/CityCard.tsx
