+++ app/ideciclo/ideciclo.tsx
@@ -1,8 +1,8 @@
 import {Layout} from "../../components/OLD/OldLayout";
-import StatisticsBox from "../../components/StatisticsBox";
+import StatisticsBox from "../contagens/components/StatisticsBox";
 import {ExplanationBox} from "../components/ExplanationBox";
 import NumberCards from "../../components/NumberCards";
-import IdecicloTable from "../../components/IdecicloTable"
+import IdecicloTable from "./components/IdecicloTable"
 
 import React, { useEffect, useState, useRef } from "react";
 /*
