+++ components/Maps/StructureMap.tsx
