+++ components/OLD/PartnersCard.tsx
