+++ servers.js
@@ -2,3 +2,6 @@ export const FEATURED_PAGES = "http://localhost:3300/featured_pages"
 export const PLATAFORM_HOME_PAGE = "https://cms.ameciclo.org/plataforma-de-dados"
 export const DOCUMENTS_PAGE = "https://cms.ameciclo.org/documentos"
 export const DOCUMENTS_DATA = "https://cms.ameciclo.org/documents"
+export const COUNTINGS_SUMMARY_DATA =`https://api.contagem.ameciclo.org/v1/cyclist-count/metadata`
+export const COUNTINGS_DATA =`https://api.contagem.ameciclo.org/v1/cyclist-count/`
+export const COUNTINGS_PAGE_DATA =`https://cms.ameciclo.org/contagens`
